Feature: Download movie files to cache, serve read operations from cache

+++ plexfuse/plex/PlexApi.py
@@ -125,7 +125,7 @@ class PlexApi:
             / self.CACHE_VERSION \
             / item.key.lstrip("/")
 
-    def download_part(self, part):
+    def download_part(self, part, overwrite=False):
         url = self.url(part.key)
         headers = {"X-Plex-Token": self.token}
         response = self.session.get(url, headers=headers, stream=True)
@@ -135,6 +135,9 @@ class PlexApi:
             raise RuntimeError(message)
 
         savepath = self.cache_path(part)
+        if overwrite is False and savepath.exists():
+            return savepath
+
         makedirs(Path(savepath).parent, exist_ok=True)
 
         with open(savepath, "wb") as handle:
